backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["xtanion"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
colorlog = "^6.9.0"
websockets = "^15.0.1"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
